アプリ化に向けてソースコードを管理する

+++ apply_lottery.py
@@ -7,16 +7,20 @@ from selenium.webdriver.support import expected_conditions as EC
 # 自作モジュール
 from scraping.ios.appium_utilities import AppiumUtilities
 from utils.spreadsheet import SpreadsheetApiClient
-from utils.gmail import extract_target_str_from_gmail_text_in_5min
 from utils.common import get_column_number_by_alphabet
+from utils.login import login_pokemon_center_online
 from config import SPREADSHEET_ID, SHEET_NAME
 
-MAX_RETRY_LOGIN = 3
+DEBUG_MODE = False
+RETRY_LOOP = 3
 MAX_RETRY_PASSCODE = 10
 
 
 def main(driver, appium_utils, user_info, top_p=1, write_col='AA'):
     """抽選応募処理"""
+
+    # スプレッドシートクライアントを定義する
+    ss = SpreadsheetApiClient()
 
     row_number = user_info["row_number"]
     email = user_info["email"]
@@ -33,82 +37,13 @@ def main(driver, appium_utils, user_info, top_p=1, write_col='AA'):
 
     try:
 
-        for retry_i in range(MAX_RETRY_LOGIN):
-
-            try:
-                # ログイン画面に遷移
-                driver.get("https://www.pokemoncenter-online.com/login/")
-                WebDriverWait(driver, 10).until(EC.presence_of_element_located((AppiumBy.TAG_NAME, "body")))
-                print("ログインページに移動しました")
-
-                time.sleep(random.uniform(3, 5))
-
-                print("IDを入力中...")
-                email_form = WebDriverWait(driver, 10).until(
-                    EC.presence_of_element_located((AppiumBy.ID, 'login-form-email'))
-                )
-                email_form.send_keys(email)
-
-                time.sleep(random.uniform(3, 5))
-
-                print("パスワードを入力中...")
-                password_form = WebDriverWait(driver, 10).until(
-                    EC.presence_of_element_located((AppiumBy.ID, 'current-password'))
-                )
-                password_form.send_keys(password)
-
-                time.sleep(random.uniform(3, 5))
-
-                print("ログインボタンをクリック中...")
-                login_button = WebDriverWait(driver, 10).until(
-                    EC.presence_of_element_located((AppiumBy.XPATH, "//*[@id='form1Button']"))
-                )
-                login_button.click()
-
-                time.sleep(15)
-                if ("ログイン" in driver.page_source and "/login/" in driver.current_url):
-                    raise Exception("ログインに失敗しました")
-
-                # 2段階認証処理
-                for retry_j in range(MAX_RETRY_PASSCODE):
-                    auth_code = extract_target_str_from_gmail_text_in_5min(
-                        to_email=email,
-                        subject_keyword="[ポケモンセンターオンライン]ログイン用パスコードのお知らせ",
-                        email_type="passcode"
-                    )
-                    if auth_code:
-                        break
-                    time.sleep(15)
-
-                print("パスコードを入力中...")
-                passcode_form = WebDriverWait(driver, 10).until(
-                    EC.presence_of_element_located((AppiumBy.ID, 'authCode'))
-                )
-                passcode_form.send_keys(auth_code)
-
-                time.sleep(random.uniform(3, 5))
-
-                print("認証ボタンをクリック中...")
-                auth_button = WebDriverWait(driver, 10).until(
-                    EC.presence_of_element_located((AppiumBy.ID, 'authBtn'))
-                )
-                auth_button.click()
-
-                time.sleep(10)
-                if ("パスコード入力" in driver.page_source and "/login-mfa/" in driver.current_url):
-                    raise Exception("2段階認証に失敗しました")
-
-                # ここまで終わったらリトライループを抜ける
-                break
-
-            except Exception as e:
-                print(f"ログイン失敗、再試行します... {e}")
-                appium_utils.delete_browser_page()
-                time.sleep(10)
-                continue
+        # ログイン処理
+        is_logged_in = login_pokemon_center_online(driver, email, password)
+        if not is_logged_in:
+            raise Exception("ログインに失敗しました")
 
         driver.get("https://www.pokemoncenter-online.com/lottery/apply.html")
-        time.sleep(random.uniform(5, 10))
+        time.sleep(random.uniform(3, 5))
 
         for index in range(top_p):
             print(f"抽選申し込み処理を開始します (index={index})")
@@ -192,19 +127,23 @@ def main(driver, appium_utils, user_info, top_p=1, write_col='AA'):
                     print("❌ 申し込みボタンのクリックに失敗しました")
                     continue
 
-                # スプレッドシートに結果を書き込む
-                ss = SpreadsheetApiClient()
-                write_col_number = get_column_number_by_alphabet(write_col) + index
-                ss.write_to_cell(
-                    spreadsheet_id=SPREADSHEET_ID,
-                    sheet_name=SHEET_NAME,
-                    row=row_number,
-                    column=write_col_number,
-                    value="応募済み"
-                )
+                # 6. 受付が完了したかチェックする
+                lottery_labels = lottery_fields = appium_utils.safe_find_elements(AppiumBy.CLASS_NAME, 'ttl', attempt=index)
+                lottery_label = lottery_labels[index]
+                print(lottery_label.get_attribute("innerText"))
+                if lottery_label.get_attribute("innerText") != "受付完了":
+                    # スプレッドシートに結果を書き込む
+                    write_col_number = get_column_number_by_alphabet(write_col) + index
+                    ss.write_to_cell(
+                        spreadsheet_id=SPREADSHEET_ID,
+                        sheet_name=SHEET_NAME,
+                        row=row_number,
+                        column=write_col_number,
+                        value="応募済み"
+                    )
 
-                time.sleep(random.uniform(10, 15))
-                print(f"🎉 抽選申し込み {index + 1} 完了!")
+                    time.sleep(random.uniform(3, 5))
+                    print(f"🎉 抽選申し込み {index + 1} 完了!")
 
             except ValueError as ve:
                 print(f"❌ {ve}")
@@ -214,53 +153,51 @@ def main(driver, appium_utils, user_info, top_p=1, write_col='AA'):
 
     except Exception as e:
         print(f"エラーが発生しました: {e}")
-        # driver.save_screenshot('error_screenshot.png')
 
     finally:
-        # # ポケセンオンラインからログアウトする
-        # print("マイページに移動中...")
-        # driver.get("https://www.pokemoncenter-online.com/mypage/")
-        # time.sleep(5)
-
-        # print("ログアウト中...")
-        # logout_buttons = appium_utils.safe_find_elements(AppiumBy.CLASS_NAME, 'logout')
-        # # if logout_buttons[0].get_attribute("innerText") == "ログアウト":
-        # if not appium_utils.safe_click(logout_buttons, 0, "ログアウト"):
-        #     print("❌ ログアウトボタンのクリックに失敗しました")
-        # time.sleep(10)
-
+        if not DEBUG_MODE:
         # ドライバーを終了
-        print("\nドライバーを終了中...")
-        appium_utils.delete_browser_page()
-        time.sleep(random.uniform(10, 15))
-        print("完了しました")
+            print("\nドライバーを終了中...")
+            appium_utils.delete_browser_page()
+            time.sleep(5)
+            print("完了しました")
+        else:
+            pass
 
 if __name__ == '__main__':
     TOP_P = 2 # 抽選申し込みを行う上位件件数
-    WRITE_COL = 'AA'  # 抽選申し込み結果を書き込む列
+    WRITE_COL = 'Z'  # 抽選申し込み結果を書き込む列
 
-    START_ROW = 27
-    END_ROW = 75
+    START_ROW = 4
+    END_ROW = 87
 
     # スプレッドシートからユーザー情報を取得する
     ss = SpreadsheetApiClient()
-    # スプレッドシートの全データをDataFrame形式で取得
-    all_data = ss.get_all_data(spreadsheet_id=SPREADSHEET_ID, sheet_name=SHEET_NAME)
-    user_info_list = ss.extract_apply_lottery_user_info(all_data, START_ROW, END_ROW, WRITE_COL)
-    print(json.dumps(user_info_list, indent=2, ensure_ascii=False))
-    print("---------------")
-    print(f"合計ユーザー数: {len(user_info_list)}")
-    print("---------------")
 
-    print("Appiumドライバーを初期化中...")
-    appium_utils = AppiumUtilities()
+    for loop in range(RETRY_LOOP):
+        print(f"{loop}回目の処理を開始します")
 
-    print("Safariを起動しました")
+        # スプレッドシートの全データをDataFrame形式で取得
+        all_data = ss.get_all_data(spreadsheet_id=SPREADSHEET_ID, sheet_name=SHEET_NAME)
+        user_info_list = ss.extract_apply_lottery_user_info(all_data, START_ROW, END_ROW, WRITE_COL, TOP_P, 'apply_lottery')
+        print(json.dumps(user_info_list, indent=2, ensure_ascii=False))
+        print("---------------")
+        print(f"合計ユーザー数: {len(user_info_list)}")
+        print("---------------")
 
-    driver = appium_utils.driver
+        print("Appiumドライバーを初期化中...")
+        appium_utils = AppiumUtilities()
 
-    for user_info in user_info_list:
-        if not user_info.get("email") or not user_info.get("password"):
-            print(f"❌ emailまたはpasswordが未設定のためスキップします: {user_info}")
-            continue
-        main(driver, appium_utils, user_info, TOP_P, WRITE_COL)
+        print("Safariを起動しました")
+
+        driver = appium_utils.driver
+
+        for user_info in user_info_list:
+            if not user_info.get("email") or not user_info.get("password"):
+                print(f"❌ emailまたはpasswordが未設定のためスキップします: {user_info}")
+                continue
+            main(driver, appium_utils, user_info, TOP_P, WRITE_COL)
+
+        # 最低3分の待機時間を確保する
+        print("次のループまで3分間待機します...")
+        time.sleep(180)
